* Added console app. * Updated web.config default tracker broadcast domain to be localhost to make it clearer to new users that this can be a domain and not just an IP.

+++ Gumshoe.IIS/Global.asax.cs
@@ -14,7 +14,7 @@ namespace Gumshoe
 			var trackerBroadcastDomain = ConfigurationManager.AppSettings ["TrackerBroadcastDomain"];
 			var trackerPort = int.Parse (ConfigurationManager.AppSettings ["TrackerPort"]);
 			var torrentsDirectory = ConfigurationManager.AppSettings ["TorrentsDirectory"];
-			var announceIntervalInSeconds = TimeSpan.FromSeconds (double.Parse (ConfigurationManager.AppSettings ["AnnounceIntervalInSeconds"]));
+			var announceIntervalInSeconds = TimeSpan.FromSeconds (int.Parse (ConfigurationManager.AppSettings ["AnnounceIntervalInSeconds"]));
 			var allowUnregisteredTorrents = bool.Parse (ConfigurationManager.AppSettings ["AllowUnregisteredTorrents"]);
 
 			// Create torrents directory.
